requirements.txt
discord.py==2.3.1
gnupg==2.3.1
python-dotenv==1.0.1
